+++ Documentation/DemoApps/GRDBAsyncDemo/GRDBAsyncDemo/Persistence.swift
@@ -7,26 +7,31 @@ extension AppDatabase {
     
     private static func makeShared() -> AppDatabase {
         do {
-            // Pick a folder for storing the SQLite database, as well as
-            // the various temporary files created during normal database
-            // operations (https://sqlite.org/tempfiles.html).
-            let fileManager = FileManager()
-            let folderURL = try fileManager
-                .url(for: .applicationSupportDirectory, in: .userDomainMask, appropriateFor: nil, create: true)
-                .appendingPathComponent("database", isDirectory: true)
-
+            // Apply recommendations from
+            // <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/databaseconnections>
+            //
+            // Create the "Application Support/Database" directory if needed
+            let fileManager = FileManager.default
+            let appSupportURL = try fileManager.url(
+                for: .applicationSupportDirectory, in: .userDomainMask,
+                appropriateFor: nil, create: true)
+            let directoryURL = appSupportURL.appendingPathComponent("Database", isDirectory: true)
+            
             // Support for tests: delete the database if requested
             if CommandLine.arguments.contains("-reset") {
-                try? fileManager.removeItem(at: folderURL)
+                try? fileManager.removeItem(at: directoryURL)
             }
             
             // Create the database folder if needed
-            try fileManager.createDirectory(at: folderURL, withIntermediateDirectories: true)
+            try fileManager.createDirectory(at: directoryURL, withIntermediateDirectories: true)
             
-            // Connect to a database on disk
-            // See https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/databaseconnections
-            let dbURL = folderURL.appendingPathComponent("db.sqlite")
-            let dbPool = try DatabasePool(path: dbURL.path)
+            // Open or create the database
+            let databaseURL = directoryURL.appendingPathComponent("db.sqlite")
+            NSLog("Database stored at \(databaseURL.path)")
+            let dbPool = try DatabasePool(
+                path: databaseURL.path,
+                // Use default AppDatabase configuration
+                configuration: AppDatabase.makeConfiguration())
             
             // Create the AppDatabase
             let appDatabase = try AppDatabase(dbPool)
@@ -59,7 +64,7 @@ extension AppDatabase {
     static func empty() -> AppDatabase {
         // Connect to an in-memory database
         // See https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/databaseconnections
-        let dbQueue = try! DatabaseQueue()
+        let dbQueue = try! DatabaseQueue(configuration: AppDatabase.makeConfiguration())
         return try! AppDatabase(dbQueue)
     }