+++ Documentation/DemoApps/GRDBAsyncDemo/GRDBAsyncDemoTests/AppDatabaseTests.swift
@@ -5,7 +5,7 @@ import GRDB
 class AppDatabaseTests: XCTestCase {
     func test_database_schema() throws {
         // Given an empty database
-        let dbQueue = try DatabaseQueue()
+        let dbQueue = try DatabaseQueue(configuration: AppDatabase.makeConfiguration())
         
         // When we instantiate an AppDatabase
         _ = try AppDatabase(dbQueue)
@@ -21,7 +21,7 @@ class AppDatabaseTests: XCTestCase {
     
     func test_savePlayer_inserts() async throws {
         // Given an empty players database
-        let dbQueue = try DatabaseQueue()
+        let dbQueue = try DatabaseQueue(configuration: AppDatabase.makeConfiguration())
         let appDatabase = try AppDatabase(dbQueue)
         
         // When we save a new player
@@ -35,7 +35,7 @@ class AppDatabaseTests: XCTestCase {
     
     func test_savePlayer_updates() async throws {
         // Given a players database that contains a player
-        let dbQueue = try DatabaseQueue()
+        let dbQueue = try DatabaseQueue(configuration: AppDatabase.makeConfiguration())
         let appDatabase = try AppDatabase(dbQueue)
         var player = try await dbQueue.write { db in
             try Player(id: nil, name: "Arthur", score: 100).inserted(db)
@@ -55,7 +55,7 @@ class AppDatabaseTests: XCTestCase {
     
     func test_deletePlayers() async throws {
         // Given a players database that contains four players
-        let dbQueue = try DatabaseQueue()
+        let dbQueue = try DatabaseQueue(configuration: AppDatabase.makeConfiguration())
         let appDatabase = try AppDatabase(dbQueue)
         let playerIds: [Int64] = try await dbQueue.write { db in
             _ = try Player(id: nil, name: "Arthur", score: 100).inserted(db)
@@ -83,7 +83,7 @@ class AppDatabaseTests: XCTestCase {
     
     func test_deleteAllPlayers() async throws {
         // Given a players database that contains players
-        let dbQueue = try DatabaseQueue()
+        let dbQueue = try DatabaseQueue(configuration: AppDatabase.makeConfiguration())
         let appDatabase = try AppDatabase(dbQueue)
         try await dbQueue.write { db in
             _ = try Player(id: nil, name: "Arthur", score: 100).inserted(db)
@@ -102,7 +102,7 @@ class AppDatabaseTests: XCTestCase {
     
     func test_refreshPlayers_populates_an_empty_database() async throws {
         // Given an empty players database
-        let dbQueue = try DatabaseQueue()
+        let dbQueue = try DatabaseQueue(configuration: AppDatabase.makeConfiguration())
         let appDatabase = try AppDatabase(dbQueue)
         
         // When we refresh players
@@ -115,7 +115,7 @@ class AppDatabaseTests: XCTestCase {
     
     func test_createRandomPlayersIfEmpty_populates_an_empty_database() throws {
         // Given an empty players database
-        let dbQueue = try DatabaseQueue()
+        let dbQueue = try DatabaseQueue(configuration: AppDatabase.makeConfiguration())
         let appDatabase = try AppDatabase(dbQueue)
         
         // When we create random players
@@ -127,7 +127,7 @@ class AppDatabaseTests: XCTestCase {
     
     func test_createRandomPlayersIfEmpty_does_not_modify_a_non_empty_database() throws {
         // Given a players database that contains one player
-        let dbQueue = try DatabaseQueue()
+        let dbQueue = try DatabaseQueue(configuration: AppDatabase.makeConfiguration())
         let appDatabase = try AppDatabase(dbQueue)
         var player = Player(id: nil, name: "Arthur", score: 100)
         try dbQueue.write { db in