+++ Documentation/DemoApps/GRDBDemoiOS/GRDBDemoiOS/AppDatabase.swift
@@ -1,11 +1,28 @@
 import GRDB
+import os.log
 
-/// AppDatabase lets the application access the database.
+/// `AppDatabase` lets the application access the database.
 ///
-/// It applies the pratices recommended at
-/// <https://github.com/groue/GRDB.swift/blob/master/Documentation/GoodPracticesForDesigningRecordTypes.md>
-final class AppDatabase {
-    /// Creates an `AppDatabase`, and make sure the database schema is ready.
+/// You create an `AppDatabase` with a connection to an SQLite database
+/// (see <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/databaseconnections>).
+///
+/// Create those connections with a configuration returned from
+/// `AppDatabase/makeConfiguration(_:)`.
+///
+/// For example:
+///
+/// ```swift
+/// // Create an in-memory AppDatabase
+/// let config = AppDatabase.makeConfiguration()
+/// let dbQueue = try DatabaseQueue(configuration: config)
+/// let appDatabase = try AppDatabase(dbQueue)
+/// ```
+struct AppDatabase {
+    /// Creates an `AppDatabase`, and makes sure the database schema
+    /// is ready.
+    ///
+    /// - important: Create the `DatabaseWriter` with a configuration
+    ///   returned by ``makeConfiguration(_:)``.
     init(_ dbWriter: any DatabaseWriter) throws {
         self.dbWriter = dbWriter
         try migrator.migrate(dbWriter)
@@ -18,18 +35,67 @@ final class AppDatabase {
     ///
     /// See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/databaseconnections>
     private let dbWriter: any DatabaseWriter
+}
+
+// MARK: - Database Configuration
+
+extension AppDatabase {
+    private static let sqlLogger = OSLog(subsystem: Bundle.main.bundleIdentifier!, category: "SQL")
     
+    /// Returns a database configuration suited for `PlayerRepository`.
+    ///
+    /// SQL statements are logged if the `SQL_TRACE` environment variable
+    /// is set.
+    ///
+    /// - parameter base: A base configuration.
+    public static func makeConfiguration(_ base: Configuration = Configuration()) -> Configuration {
+        var config = base
+        
+        // An opportunity to add required custom SQL functions or
+        // collations, if needed:
+        // config.prepareDatabase { db in
+        //     db.add(function: ...)
+        // }
+        
+        // Log SQL statements if the `SQL_TRACE` environment variable is set.
+        // See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/database/trace(options:_:)>
+        if ProcessInfo.processInfo.environment["SQL_TRACE"] != nil {
+            config.prepareDatabase { db in
+                db.trace {
+                    // It's ok to log statements publicly. Sensitive
+                    // information (statement arguments) are not logged
+                    // unless config.publicStatementArguments is set
+                    // (see below).
+                    os_log("%{public}@", log: sqlLogger, type: .debug, String(describing: $0))
+                }
+            }
+        }
+        
+#if DEBUG
+        // Protect sensitive information by enabling verbose debugging in
+        // DEBUG builds only.
+        // See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/configuration/publicstatementarguments>
+        config.publicStatementArguments = true
+#endif
+        
+        return config
+    }
+}
+
+// MARK: - Database Migrations
+
+extension AppDatabase {
     /// The DatabaseMigrator that defines the database schema.
     ///
     /// See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/migrations>
     private var migrator: DatabaseMigrator {
         var migrator = DatabaseMigrator()
         
-        #if DEBUG
+#if DEBUG
         // Speed up development by nuking the database when migrations change
         // See <https://swiftpackageindex.com/groue/grdb.swift/documentation/grdb/migrations>
         migrator.eraseDatabaseOnSchemaChange = true
-        #endif
+#endif
         
         migrator.registerMigration("createPlayer") { db in
             // Create a table
@@ -51,6 +117,7 @@ final class AppDatabase {
 }
 
 // MARK: - Database Access: Writes
+// The write methods execute invariant-preserving database transactions.
 
 extension AppDatabase {
     /// Saves (inserts or updates) a player. When the method returns, the