+++ Tests/GRDBTests/ColumnExpressionTests.swift
@@ -20,7 +20,7 @@ class ColumnExpressionTests: GRDBTestCase {
             }
             
             // Test databaseSelection
-            static let databaseSelection: [SQLSelectable] = [Columns.id, Columns.name, Columns.score]
+            static let databaseSelection: [any SQLSelectable] = [Columns.id, Columns.name, Columns.score]
             
             init(row: Row) {
                 // Test row subscript
@@ -80,7 +80,7 @@ class ColumnExpressionTests: GRDBTestCase {
             }
             
             // Test databaseSelection
-            static let databaseSelection: [SQLSelectable] = [Columns.id, Columns.name, Columns.score]
+            static let databaseSelection: [any SQLSelectable] = [Columns.id, Columns.name, Columns.score]
             
             init(row: Row) {
                 // Test row subscript
@@ -148,7 +148,7 @@ class ColumnExpressionTests: GRDBTestCase {
             }
             
             // Test databaseSelection
-            static let databaseSelection: [SQLSelectable] = [Columns.id, Columns.name, Columns.score]
+            static let databaseSelection: [any SQLSelectable] = [Columns.id, Columns.name, Columns.score]
             
             static var testRequest: QueryInterfaceRequest<Player> {
                 // Test expression derivation
@@ -196,7 +196,7 @@ class ColumnExpressionTests: GRDBTestCase {
             }
             
             // Test databaseSelection
-            static let databaseSelection: [SQLSelectable] = [CodingKeys.id, CodingKeys.name, CodingKeys.score]
+            static let databaseSelection: [any SQLSelectable] = [CodingKeys.id, CodingKeys.name, CodingKeys.score]
             
             static var testRequest: QueryInterfaceRequest<Player> {
                 // Test expression derivation