+++ Tests/GRDBTests/FoundationUUIDTests.swift
@@ -3,7 +3,7 @@ import Foundation
 import GRDB
 
 class FoundationUUIDTests: GRDBTestCase {
-    private func assert(_ value: DatabaseValueConvertible?, isDecodedAs expectedUUID: UUID?) throws {
+    private func assert(_ value: (any DatabaseValueConvertible)?, isDecodedAs expectedUUID: UUID?) throws {
         try makeDatabaseQueue().read { db in
             if let expectedUUID {
                 let decodedUUID = try UUID.fetchOne(db, sql: "SELECT ?", arguments: [value])